[DH] improve Fel Devastation handling
