+++ src/analysis/retail/demonhunter/vengeance/CombatLogParser.tsx
@@ -59,10 +59,11 @@ import FieryBrand2 from './modules/core/MajorDefensives/FieryBrand';
 import DemonSpikes from './modules/core/MajorDefensives/DemonSpikes';
 import Metamorphosis from './modules/core/MajorDefensives/Metamorphosis';
 import Guide from './Guide';
-import FelDevastationNormalizer from './normalizers/FelDevastationNormalizer';
+import FelDevastationLinkNormalizer from './normalizers/FelDevastationLinkNormalizer';
 import CycleOfBindingNormalizer from './normalizers/CycleOfBindingNormalizer';
 import CycleOfBinding from './modules/talents/CycleOfBinding';
 import ConsumingFireNormalizer from '../shared/normalizers/ConsumingFireNormalizer';
+import FelDevastationNormalizer from './normalizers/FelDevastationNormalizer';
 
 class CombatLogParser extends CoreCombatLogParser {
   static specModules = {
@@ -91,6 +92,7 @@ class CombatLogParser extends CoreCombatLogParser {
     defensiveBuffLinkNormalizer: DefensiveBuffLinkNormalizer,
     theHuntNormalizer: TheHuntNormalizer,
     felDevastationNormalizer: FelDevastationNormalizer,
+    felDevastationLinkNormalizer: FelDevastationLinkNormalizer,
     sigilOfFlamesNormalizer: SigilOfFlameNormalizer,
     cycleOfBindingNormalizer: CycleOfBindingNormalizer,
     consumingFireNormalizer: ConsumingFireNormalizer,
